pyproject.toml
[project]
name = "crypto-dashboard"
version = "0.1.0"
description = "Real-time Crypto API using FastAPI"
authors = [
    {name = "leblancr",email = "rich.leblanc1@pm.me"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "uvicorn (>=0.37.0,<0.38.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"
mypy = "^1.18.2"
pytest = "^8.4.2"

